backend/requirements.txt
# Core web/API – pinned where stable
fastapi==0.129.0
uvicorn[standard]==0.34.0
pydantic>=2.7.0,<3.0.0                # FastAPI requires >=2.7

# File & image handling
python-multipart==0.0.20
pillow==11.1.0

# Supabase client – use latest (fixes your invalid 2.2.28 pin)
supabase==2.28.0

# Database driver
psycopg[binary]
asyncpg==0.29.0                        # Added for async PostgreSQL connection

# Authentication & JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt>=2.10.1
authlib==1.6.8

# LangChain & related (your versions are ok, but consider upgrading later)
langchain==1.2.10
langchain-community==0.4.1
langchain-core==1.2.13

# Embeddings & models
transformers==4.44.0
sentence-transformers==3.3.0
numpy==2.2.0

# Utils
requests==2.32.5
python-dotenv==1.0.1

# Vector store – 1.5.0 should work; bump if build fails
chromadb==1.5.0                      # or ==1.5.1 / >=1.5.1 if needed

# Optional / commented out for now
google-genai==1.64.0
# ollama==0.6.1